requirements.txt
torch>=2.0.0
transformers>=4.46.0
datasets>=2.14.0
accelerate>=0.24.0
flash-attn>=2.3.0
xformers>=0.0.22
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.8
pillow>=9.5.0
tqdm>=4.65.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
flask>=3.0.0
pytesseract>=0.3.10
